+++ src/ts/Parsers/PromoCampaignParser.ts
@@ -10,6 +10,7 @@ import { Session } from 'Models/Session';
 import { HTML } from 'Models/Assets/HTML';
 
 export class PromoCampaignParser extends CampaignParser {
+    public static ContentType = 'purchasing/iap';
     public parse(nativeBridge: NativeBridge, request: Request, response: AuctionResponse, session: Session, gamerId: string, abGroup: number): Promise<Campaign> {
         const promoJson = JsonParser.parse(response.getContent());
         if (promoJson && promoJson.iapProductId) {